django_project/pyproject.toml
[tool.poetry]
name = "django-project"
version = "0.1.0"
description = ""
authors = ["Wargoth_ks <121107554+Wargoth-ks@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
jinja2 = "^3.1.2"
psycopg2 = "^2.9.9"
pymongo = "^4.5.0"
pillow = "^10.0.1"
tqdm = "^4.66.1"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
